+++ src/index.ts
@@ -2,6 +2,4 @@ export * from './association';
 export * from './attribute';
 export * from './metadata';
 export * from './model';
-export * from './property';
-export * from './safe';
 export * from './validation';